requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
tqdm>=4.65.0
opencv-python>=4.7.0

# Data handling
h5py>=3.8.0
Pillow>=9.0.0
plyfile>=0.7.4

# Configuration and logging
pyyaml>=6.0
wandb>=0.13.0

# Optional but recommended
tensorboard>=2.11.0
jupyterlab>=3.5.0

# For testing
pytest>=7.2.0
pytest-cov>=4.0.0